requirements.txt
Jinja2==2.10.3
PyInstaller==3.5
PyYAML==5.2
termcolor==1.1.0